refactor(attention): move ffpa_attention backend to components/attention
